+++ product_portfolio/views.py
@@ -1961,40 +1961,15 @@ class PullProjectDataFromScanCodeIOView(
     def get_success_url(self):
         return f"{self.object.get_absolute_url()}#imports"
 
-    def get_project_data(self, project_name_or_uuid):
-        scancodeio = ScanCodeIO(self.request.user)
-        for field_name in ["name", "uuid"]:
-            project_data = scancodeio.find_project(**{field_name: project_name_or_uuid})
-            if project_data:
-                return project_data
-
     def form_valid(self, form):
-        project_name_or_uuid = form.cleaned_data.get("project_name_or_uuid")
-        project_data = self.get_project_data(project_name_or_uuid)
+        self.object = self.get_object()
 
-        if not project_data:
-            msg = f'Project "{project_name_or_uuid}" not found on ScanCode.io.'
-            messages.error(self.request, msg)
+        try:
+            form.submit(product=self.object, user=self.request.user)
+        except ValidationError as error:
+            messages.error(self.request, error)
             return redirect(self.object.get_absolute_url())
 
-        scancode_project = ScanCodeProject.objects.create(
-            product=self.object,
-            dataspace=self.object.dataspace,
-            type=ScanCodeProject.ProjectType.PULL_FROM_SCANCODEIO,
-            project_uuid=project_data.get("uuid"),
-            update_existing_packages=form.cleaned_data.get("update_existing_packages"),
-            scan_all_packages=False,
-            status=ScanCodeProject.Status.SUBMITTED,
-            created_by=self.request.user,
-        )
-
-        transaction.on_commit(
-            lambda: tasks.pull_project_data_from_scancodeio.delay(
-                scancodeproject_uuid=scancode_project.uuid,
-            )
-        )
-
-        project_name = project_data.get("name")
-        msg = f'Packages import from ScanCode.io "{project_name}" in progress...'
+        msg = "Packages import from ScanCode.io in progress..."
         messages.success(self.request, msg)
         return super().form_valid(form)